api: simplify archive policy validation
The current validation code validates measure aggregates retrieval against the
whole list of valid aggregation methods. That serves little purpose as it is
again validated just after against the actual list of existing aggregation
method defined in the archive policy attached to this/these metric(s).

This patch removes that validation and adds a missing test for the aggregation
endpoint.