improve documentation on tuning the number of sacks
The documentation suggested that the number of sacks could be tuned
via a configuration option the in gnocchi.conf file.  This commit
corrects the docs to indicate that the value can be changed only with
the gnocchi-change-sack-size command.

This commit also corrects spelling, grammar, and formatting issues in
the same section.