storage: do not return empty data set on non-existing metric
This decision is now available for the upper layer.

This patch also fixes the S3 driver which was not checking properly for metric
existence. Listing objects never returns a NoSuchKey error, at worse it returns
an empty content list. We therefore need an extra check with a HEAD request on
the unaggregated timeseries to be sure that the metric exists. It therefore
needs to be created at _get_or_create_unaggregated_timeseries_unbatched() like
the other drivers do.

This was not caught earlier because of the return empty by default mechanism,
but it is correctly tested.