storage: stop raising AggregationDoesNotExist on measures retrieval
Raising an exception at this point is damageable as when multiple SplitKeys are
being fetched at the same time, only one being absent will cancel the call. The
caller is already solid and know show to handle None as a return value so just
use None to indicate the split is non-existing.