swift: Allow to configure the requests pool size
By default, python-requests share a pool for the whole process and this
one have only 10 parallels connections. This change use the
keystoneauth1 adapter. So, we have a pool dedicated for swift and we can
configure it. And we also get all tcp keepalive workaround for free.

This allows keep the default for 'retries'.

Closes: #509