Fixup rendering on Pypi
The description of the project is currently not rendered on
Pypi.  Instead, Pypi shows raw rst.  Checking the page with
"readme_renderer" reports "undefined substitutions" for "time
series" and "aggregates".

This patch adds a before_deploy section to .travis.yml that
uses sed to remove matching pairs of pipes from the README.

The resulting README was tested again with the readme renderer.

See Also:
    https://pypi.org/project/readme_renderer/