Fix compatibility with MySQL >= 8
Under newer MySQL versions check constraints are supported and
must have unique names.

Drop and re-create the check constraints on the resource and
resource_history tables to ensure that they are uniquely named
for upgrades in existing deployments.

Ensure that check constraints are created with unique names in
new deployments.