Set json template for DELETE requests
Setting expose to json will cause content-type
internally for Pecan to be set to application/json.

If this is not set Pecan will throw a NotAcceptable
error if it cannot parse or understand what type
of request it is.

This does not affect the content-type in the response
for DELETE calls.