Remove Travis CI functional tests
This removes the functional testing that was
used in Travis CI.

This still leaves the deploy pipeline and fixes
the before_deploy script so that python3 is installed
so the pypi upload works.

It adds the "os" key to the Travis CI config which was
a warning. The "user" key in the deploy pipeline is deprecated
and should be "username".