When we use python3, it still find python2
The args[] use "python", and if the version of python is "3",
it will can't find python, because the name is "python3" in o

Add a judge of python version, if the version of python is 3,
set plugin to "python3", if not, plugin still "python".

If the version of Python is 3, api.py can run.

(cherry picked from commit 325628c2a813c53d3036e661a0e7c281377d5059)