Drop use of setuptools extension
We were wrapping easy_install's 'install_scripts' and 'develop' methods
to generate a custom WSGI script. This was a replacement of
functionality previously provided by PBR. Unfortunately easy_install is
deprecated for removal at a future date and we don't yet have a modern
equivalent in pip [1]. As such, we need a migration plan for this and it
seems like the old school 'scripts' option is probably it. The main
downside of this is that the shebang isn't updated accordingly, which
means this won't work very well in virtualenvs, but I'm hoping that the
use of '/usr/bin/env' will allow us to work around this. At least in
local testing, it seems like this is the case:

  ❯ virtualenv .venv
  ❯ pip install .
  ❯ cat .venv/bin/gnocchi-api | head -1
  #!/gnocchi/.venv/bin/python

[1] https://discuss.python.org/t/adding-support-for-wsgi-scripts-entrypoint/30905

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>