redis: Fix wrong types of some options
The socket_keepalive option in RedisClient is not an integer but
a boolean. Also the healthcheck_interval option can accept not only
integers but also float values. This fixes the types to parse these
options accordingly.

Note this change still keeps support for socket_keepalive=1 or
socket_keepalive=0 , but would remove support for the other integer
values such as 2 or -1.