Improve documentation for aggregates API and the `use_history` option
The `use_history` option for the aggregates API allows one to retrieve
the exact attributes of a resource that were present during a given
timestamp for a measurement obtained via the aggregates API. However,
this API has a downside, which is the inability to execute
"re-aggregations" for now. Therefore, if one needs to execute there
"re-aggregation" of computed measures, she/he will need to use the
aggregates API with the option `use_history` as `False`.

This patch makes that clear for the end users.