Optimize Gnocchi aggregates API (#1307)
* Optimize Gnocchi aggregates API

Gnocchi aggregates API can present slowness with time, due to the growing number of resources and revisions in large scale OpenStack Cloud environments. This happens due to the situation where Gnocchi does not apply filters in the MySQL queries in the resource tables. I mean, start/stop filters are not applied, and others as well, which makes Gnocchi to manipulate all dataset in the MySQL database and then the filtering is executed in Python.

To cope with that situation, we are proposing an optimization to improve the query executed by Gnocchi in MySQL, and apply the filtering right away; thus, the dataset manipulated by MySQL will be drastically reduced. After applying the patch, we noticed a reduction of 5-6 folds in teh response time for the Gnocchi aggregates API.

* Use proper start timeframe when handling different resources together

* address Daniel's review

* Add some extra information regarding re-aggregation with the use of use_history

* Chungg review

* revert a change that does not work

* Remove documentation change