Clean raw data metrics that do not receive measures (#1385)
The truncating of the raw metrics data is only done when new measures
are pushed. Therefore, if no new measures are pushed, and the archive
policy is updated to reduce the back window, the raw data points for
metrics that are not receiving new data points are never truncated.

The goal of this PR is to propose a method to identify metrics that are
in "inactive state", meaning, not receiving new data points, and execute
their raw data points truncation when the archive policy backwindow is
changed.