Remove the start/end filtering from aggregates resource query
The use of the start/end filtering in resources listage in the
aggregates api was generating an undesired behavior that was
not showing measures of resources when the measures timestamp
were from before the resources start time, which were displayed
before this filter was introduced. This patch restores the old
behavior.