Increase original resource ID max length to 289 chars
Ceilometer formats Swift container resource IDs in the format
`<project ID>_<container name>`, where the project ID is 32 characters
and container names can be up to 256 characters long, resulting in a
resource ID with max length 289 characters.

To accommodate this, expand the `original_resource_id` column
to have a max length of 289 characters.