Swift: Add a config option for prefixing all containers
The `swift_container_prefix` option by itself only prefixes
the containers created for metric storage.

This is not enough to allow multiple Gnocchi deployments to share
the same OpenStack project, as additional containers are created
for Gnocchi configuration as well as for incoming metric storage,
if Swift is also used as the incoming storage driver.

This adds a new `swift_container_prefix_all` flag for enabling
prepending `swift_container_prefix` to **all** Swift containers
that can be made by Gnocchi.

This allows new deployments of Gnocchi to be configured in a way
that multiple deployments can share the project (e.g. a service
project in a multi-region cloud). Unfortunately, existing
Gnocchi deployments cannot easily be reconfigured to enable
prefixes, nor can the prefix easily be changed.

Fixes #1067.