Run metricd more frequently in functional tests
The metricd thread in the functional tests waits for 100ms between runs.

This is actually quite a long time to wait, as the functional tests can run much faster than that. This results in a race condition where earlier tests submit measures, and they are not actually processed by the time the test that needs them is run, and fails because they are not done (and there's no waits or retries or anything).

Tighten up the metricd thread loop so that it waits only 10ms between runs. This will increase load but the more frequent runs should be good enough for the tests to pass.