Gnocchi deleted resources cleanup
After the development of improvements [1] and [2], we move with this
issue to the last step of the improvement work regarding resource's
metadata management in Gnocchi.

The idea is to create an option in Gnocchi to automatically clean up
deleted resources after a given period of time (configured by the
operator). For instance, the idea is that somebody can configure
Gnocchi to automatically remove resources marked as deleted after 12
months or 24 months, and so on.

Therefore, with this feature, Gnocchi would be able to auto-clean
itself; thus, this can have a good impact on the aggregates API, which
manipulates the "resource" table.

[1] https://github.com/gnocchixyz/gnocchi/pull/1386
[2] https://github.com/gnocchixyz/gnocchi/pull/1385