Fix ended_at field normalization
When Gnocchi is executing the "ended_at" field normalization, which is
executed if all metrics of a resource stops receiving measures, there is
a bug that will cause Gnocchi to not generate properly the resource
revisions, which in turn can affect the aggregates API, when filtering
by resource type attributes.

This PR introduces the code changes to handle this issue.