Define ResourceHistory.metrics relationship overlap
Set `overlaps="metrics,resource"` on the `ResourceHistory.metrics` relationship defining a one-to-many relationship between resource history columns and metrics mapped to the resource.

This lets SQLAlchemy know to expect an overlap for the relationships between `Metric`, `Resource` and `ResourceHistory`, suppressing a warning output when the table schema is loaded. Because `back_populates` is configured to go between `Metric` and `Resource`, `ResourceHistory` rows should not be returned when `Metric.resource` is used.

Fixes #1406.