Support returning boolean aggregates
Some aggregates operations (e.g. `=`) return a `numpy.bool` object, and it is possible to get Gnocchi to return these, but the JSON encoder fails to encode them correctly.

This could be useful for some aggregations (e.g. checking status of OpenStack resources is a specific value), so add support to `gnocchi.json.to_primitive` for encoding `numpy.bool` to a floating-point number so it can be returned in aggregates queries as a `numbool`-style value.

Fixes #1473.